+++ src/decorators/children/test.js
@@ -38,8 +38,8 @@ const makeDatabase = () =>
         ],
       }),
     },
-
     modelClasses: [MockParent, MockChild],
+    actionsEnabled: true,
   })
 
 describe('decorators/children', () => {
@@ -47,7 +47,7 @@ describe('decorators/children', () => {
     const database = makeDatabase()
     database.adapter.batch = jest.fn()
 
-    const parentMock = await database.collections.get('mock_parent').create()
+    const parentMock = await database.action(() => database.collections.get('mock_parent').create())
 
     const expectedQuery = database.collections
       .get('mock_child')
@@ -58,7 +58,7 @@ describe('decorators/children', () => {
     const database = makeDatabase()
     database.adapter.batch = jest.fn()
 
-    const parent = await database.collections.get('mock_parent').create()
+    const parent = await database.action(() => database.collections.get('mock_parent').create())
     class ParentProxy {
       asModel = parent
 
@@ -71,7 +71,7 @@ describe('decorators/children', () => {
     const database = makeDatabase()
     database.adapter.batch = jest.fn()
 
-    const parent = await database.collections.get('mock_parent').create()
+    const parent = await database.action(() => database.collections.get('mock_parent').create())
 
     const spy = jest.spyOn(logger, 'error').mockImplementation(() => {})
     parent.children = []